Fix for arrow keys being reported as unidentified.
And minor improvements
