+++ Cargo.toml
@@ -31,6 +31,7 @@ log = "0.4"
 serde = { version = "1", optional = true, features = ["serde_derive"] }
 raw-window-handle = "0.3"
 bitflags = "1"
+nameof = "1"
 
 [dev-dependencies]
 image = "0.23.12"
